+++ internal/provider/data_source_zone.go
@@ -3,9 +3,9 @@ package provider
 import (
 	"context"
 	"fmt"
-	"log"
 
 	"github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -63,7 +63,7 @@ func dataSourceCloudflareZone() *schema.Resource {
 }
 
 func dataSourceCloudflareZoneRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	log.Printf("[DEBUG] Reading Zones")
+	tflog.Debug(ctx, fmt.Sprintf("Reading Zones"))
 	client := meta.(*cloudflare.API)
 	zoneID := d.Get("zone_id").(string)
 	name := d.Get("name").(string)