+++ internal/provider/provider_test.go
@@ -1,11 +1,13 @@
 package provider
 
 import (
+	"context"
 	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 const (
@@ -14,6 +16,20 @@ const (
 )
 
 var (
+	// testAccProvider is the "main" provider instance.
+	//
+	// This Provider can be used in testing code for API calls without requiring
+	// the use of saving and referencing specific ProviderFactories instances.
+	//
+	// testAccPreCheck(t) must be called before using this provider instance.
+	testAccProvider *schema.Provider
+
+	// providerFactories are used to instantiate a provider during acceptance
+	// testing. The factory function will be invoked for every Terraform CLI
+	// command executed to create a provider server to which the CLI can
+	// reattach.
+	providerFactories map[string]func() (*schema.Provider, error)
+
 	// Integration test account ID.
 	testAccCloudflareAccountID string = "f037e56e89293a057740de681ac9abbe"
 
@@ -26,13 +42,19 @@ var (
 	testAccCloudflareAltZoneID string = "b72110c08e3382597095c29ba7e661ea"
 	// Integration test account alternate zone name.
 	testAccCloudflareAltZoneName string = "terraform2.cfapi.net"
+)
 
+func init() {
+	testAccProvider = New("dev")()
 	providerFactories = map[string]func() (*schema.Provider, error){
 		"cloudflare": func() (*schema.Provider, error) {
 			return New("dev")(), nil
 		},
 	}
-)
+}
+func TestProvider_impl(t *testing.T) {
+	var _ *schema.Provider = New("dev")()
+}
 
 func TestProvider(t *testing.T) {
 	if err := New("dev")().InternalValidate(); err != nil {
@@ -49,6 +71,11 @@ func testAccPreCheck(t *testing.T) {
 
 	if v := os.Getenv("CLOUDFLARE_ZONE_ID"); v == "" {
 		t.Fatal("CLOUDFLARE_ZONE_ID must be set for this acceptance test")
+	}
+
+	err := testAccProvider.Configure(context.Background(), terraform.NewResourceConfigRaw(nil))
+	if err != nil {
+		t.Fatal(err)
 	}
 }