+++ internal/provider/resource_cloudflare_access_application_test.go
@@ -496,7 +496,7 @@ resource "cloudflare_access_application" "%[1]s" {
 }
 
 func testAccCheckCloudflareAccessApplicationDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_access_application" {