+++ internal/provider/resource_cloudflare_access_bookmark_test.go
@@ -80,7 +80,7 @@ resource "cloudflare_access_bookmark" "%[1]s" {
 }
 
 func testAccCheckCloudflareAccessBookmarkDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_access_bookmark" {