+++ internal/provider/resource_cloudflare_access_service_tokens_test.go
@@ -265,7 +265,7 @@ resource "cloudflare_access_service_token" "%[1]s" {
 }
 
 func testAccCheckCloudflareAccessServiceTokenDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_access_service_token" {