+++ internal/provider/resource_cloudflare_teams_location_test.go
@@ -54,7 +54,7 @@ resource "cloudflare_teams_location" "%[1]s" {
 }
 
 func testAccCheckCloudflareTeamsLocationDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_teams_location" {