+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 3.16.0 (Unreleased)
 
+NOTES:
+
+* provider: swap internal logging mechanism to use `tflog` ([#1638](https://github.com/cloudflare/terraform-provider-cloudflare/issues/1638))
+* provider: updated internal package structure of repository ([#1636](https://github.com/cloudflare/terraform-provider-cloudflare/issues/1636))
+
 ENHANCEMENTS:
 
 * resource/cloudflare_logpush_job: add support for specifying `frequency` ([#1634](https://github.com/cloudflare/terraform-provider-cloudflare/issues/1634))